fix(striker-ui): isolate EditFenceDeviceInputGroup
